refactor: flatten hook imports
