+++ packages/design-system/src/hooks/use-defaults/index.test.tsx
@@ -12,8 +12,8 @@
 
 import { render, renderHook, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
+import { DefaultsProvider, useDefaults } from './';
 import type { DefaultsContext } from './types';
-import { DefaultsProvider, useDefaults } from './use-defaults';
 
 describe('DefaultsProvider', () => {
   it('should render', () => {
