+++ packages/design-system/src/hooks/use-theme/index.test.tsx
@@ -12,8 +12,8 @@
 
 import { render, renderHook, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
+import { ThemeProvider, useTheme } from './';
 import type { ThemeContext } from './types';
-import { ThemeProvider, useTheme } from './use-theme';
 
 describe('ThemeProvider', () => {
   it('should render', () => {
