+++ packages/design-system/src/hooks/use-theme/use-theme.stories.tsx
@@ -1,6 +1,22 @@
+/*
+ * Copyright 2025 Hypergiant Galactic Systems Inc. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
 import type { Story, StoryDefault } from '@ladle/react';
-import { genericColorVars, semanticColorVars, sizeVars } from '../../styles';
-import { useTheme } from './use-theme';
+import { useTheme } from '.';
+import {
+  genericColorVars,
+  semanticColorVars,
+  sizeVars,
+} from '../../styles/theme.css';
 
 export default {
   title: 'Hooks',
