+++ packages/design-system/src/hooks/use-tree/index.test.ts
@@ -12,8 +12,8 @@
 
 import { act, renderHook } from '@testing-library/react';
 import { describe, expect, it, vi } from 'vitest';
-import type { TreeGroupNode, UseTreeOptions } from '../../types';
-import { useTree } from './use-tree';
+import type { TreeGroupNode, UseTreeOptions } from '../../types/use-tree';
+import { useTree } from './';
 
 function setup({
   nodes = [
